requirements.txt
fastapi[all]==0.112.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==1.4.0